@@ -2,28 +2,30 @@ package com.example.spacenbeyond.view;
 
 import android.app.DatePickerDialog;
 import android.content.Context;
-import android.graphics.Color;
+import android.content.Intent;
+import android.graphics.Bitmap;
+import android.graphics.drawable.BitmapDrawable;
+import android.net.Uri;
+import android.os.Build;
 import android.os.Bundle;
+import android.os.Environment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.DatePicker;
 import android.widget.ImageView;
 import android.widget.ProgressBar;
-import android.widget.ScrollView;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
 import androidx.lifecycle.ViewModelProviders;
-
 import com.example.spacenbeyond.R;
+import com.example.spacenbeyond.model.PhotoEntity;
+import com.example.spacenbeyond.model.PhotoResponse;
 import com.example.spacenbeyond.viewmodel.PhotoViewModel;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.button.MaterialButton;
 import com.google.firebase.ml.common.modeldownload.FirebaseModelDownloadConditions;
 import com.google.firebase.ml.naturallanguage.FirebaseNaturalLanguage;
@@ -31,42 +33,46 @@ import com.google.firebase.ml.naturallanguage.translate.FirebaseTranslateLanguag
 import com.google.firebase.ml.naturallanguage.translate.FirebaseTranslator;
 import com.google.firebase.ml.naturallanguage.translate.FirebaseTranslatorOptions;
 import com.squareup.picasso.Picasso;
-import com.vivekkaushik.datepicker.DatePickerTimeline;
-import com.vivekkaushik.datepicker.OnDateSelectedListener;
-
+import java.io.ByteArrayOutputStream;
+import java.io.File;
+import java.io.FileOutputStream;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
+import static com.example.spacenbeyond.util.AppUtil.verificaConexaoComInternet;
 
 public class HomeFragment extends Fragment {
 
-    private DatePickerTimeline datePickerTimeline;
+    private PhotoResponse photoResponse;
+
+    private TextView textViewDia;
     private TextView textViewMes;
     private TextView textViewAno;
 
     private ImageView imageViewCalendar;
     private ImageView imageViewUser;
 
-    private FragmentManager manager;
-
     private ProgressBar progressBar;
     private PhotoViewModel photoViewModel;
     private TextView textViewFoto;
     private TextView textViewAutor;
     private ImageView imageViewFoto;
+    private ImageView imageFavorite;
+    private ImageView imageShare;
     private TextView textViewDescricao;
     private MaterialButton materialButtonPT;
     private MaterialButton materialButtonENG;
-    private ScrollView container;
-    public String dateRequest;
+    private String dateRequest;
     public static final String API_KEY = "xePvTtG6Ef3It4NuYb1mdPnKkRTFXnAAmgk0taFl";
 
     private Context context;
 
-    public HomeFragment(){ }
+    public HomeFragment(){
+
+    }
 
     @Override
-    public void onAttach(Context context){
+    public void onAttach(@NonNull Context context){
         this.context = context;
         super.onAttach(context);
     }
@@ -78,83 +84,106 @@ public class HomeFragment extends Fragment {
         initViews(view);
 
         final Calendar myCalendar = Calendar.getInstance();
-        final DatePickerDialog.OnDateSetListener date = new DatePickerDialog.OnDateSetListener() {
+        final DatePickerDialog.OnDateSetListener date = (view1, year, monthOfYear, dayOfMonth) -> {
 
-            @Override
-            public void onDateSet(DatePicker view, int year, int monthOfYear,
-                                  int dayOfMonth) {
-                // TODO Auto-generated method stub
-                myCalendar.set(Calendar.YEAR, year);
-                myCalendar.set(Calendar.MONTH, monthOfYear);
-                myCalendar.set(Calendar.DAY_OF_MONTH, dayOfMonth);
-                updateLabel(myCalendar);
+            myCalendar.set(Calendar.YEAR, year);
+            myCalendar.set(Calendar.MONTH, monthOfYear);
+            myCalendar.set(Calendar.DAY_OF_MONTH, dayOfMonth);
+            updateLabel(myCalendar);
 
-                String mes = "";
-                if (monthOfYear == 1) {
-                    mes = "01";
-                }
-                else if (monthOfYear == 2) {
-                    mes = "02";
-                }
-                else if (monthOfYear == 3) {
-                    mes = "03";
-                }
-                else if (monthOfYear == 4) {
-                    mes = "04";
-                }
-                else if (monthOfYear == 5) {
-                    mes = "05";
-                }
-                else if (monthOfYear == 6) {
-                    mes = "06";
-                }
-                else if (monthOfYear == 7) {
-                    mes = "07";
-                }
-                else if (monthOfYear == 8) {
-                    mes = "08";
-                }
-                else if (monthOfYear == 9) {
-                    mes = "09";
-                }
-                dateRequest = year + "/" + mes + "/" + dayOfMonth;
-                getPhotoOfDay();
+            String mes = "";
+            if (monthOfYear == 1) {
+                mes = "01";
             }
+            else if (monthOfYear == 2) {
+                mes = "02";
+            }
+            else if (monthOfYear == 3) {
+                mes = "03";
+            }
+            else if (monthOfYear == 4) {
+                mes = "04";
+            }
+            else if (monthOfYear == 5) {
+                mes = "05";
+            }
+            else if (monthOfYear == 6) {
+                mes = "06";
+            }
+            else if (monthOfYear == 7) {
+                mes = "07";
+            }
+            else if (monthOfYear == 8) {
+                mes = "08";
+            }
+            else if (monthOfYear == 9) {
+                mes = "09";
+            }
+            dateRequest = year + "-" + mes + "-" + dayOfMonth;
+            getPhotoOfDay();
         };
 
-        imageViewCalendar.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                new DatePickerDialog(context, date, myCalendar .get(Calendar.YEAR), myCalendar.get(Calendar.MONTH), myCalendar.get(Calendar.DAY_OF_MONTH)).show();
-            }
+        imageViewCalendar.setOnClickListener(v -> new DatePickerDialog(context, date, myCalendar .get(Calendar.YEAR), myCalendar.get(Calendar.MONTH), myCalendar.get(Calendar.DAY_OF_MONTH)).show());
+        imageViewUser.setOnClickListener(v -> replaceFragments(new EditContaFragment()));
+
+        materialButtonPT.setOnClickListener(v -> {
+            progressBar.setVisibility(View.VISIBLE);
+            getPortugueseTitle();
+            getPortugueseDescription();
+            progressBar.setVisibility(View.GONE);
         });
-        imageViewUser.setOnClickListener(new View.OnClickListener() {
+        materialButtonENG.setOnClickListener(v -> {
+            progressBar.setVisibility(View.VISIBLE);
+            getEnglishTitle();
+            getEnglishDescription();
+            progressBar.setVisibility(View.GONE);
+        });
+
+        imageFavorite.setOnClickListener(new View.OnClickListener() {
+
             @Override
-            public void onClick(View v) {
-                replaceFragments(R.id.frameContainer, new EditContaFragment());
+            public void onClick(View view) {
+
+                if (verificaConexaoComInternet(getContext())) {
+                    photoViewModel.salvarFavorito(photoResponse);
+                }
+                else {
+                    PhotoEntity photoEntity = new PhotoEntity(photoResponse.getCopyright(), photoResponse.getDate(), photoResponse.getExplanation(), photoResponse.getTitle(), photoResponse.getUrl());
+                    photoViewModel.insereDadosBd(photoEntity);
+                }
             }
         });
 
-        materialButtonPT.setOnClickListener(new View.OnClickListener() {
+        imageShare.setOnClickListener(new View.OnClickListener() {
+            @RequiresApi(api = Build.VERSION_CODES.O)
             @Override
             public void onClick(View v) {
-                progressBar.setVisibility(View.VISIBLE);
-                getPortugueseTitle();
-                getPortugueseDescription();
-                progressBar.setVisibility(View.GONE);
-            }
-        });
-        materialButtonENG.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progressBar.setVisibility(View.VISIBLE);
-                getEnglishTitle();
-                getEnglishDescription();
-                progressBar.setVisibility(View.GONE);
+
+                try {
+                    BitmapDrawable drawable = (BitmapDrawable) imageViewFoto.getDrawable();
+                    Bitmap bitmap = drawable.getBitmap();
+
+                    ByteArrayOutputStream bytes = new ByteArrayOutputStream();
+                    bitmap.compress(Bitmap.CompressFormat.JPEG, 100, bytes);
+                    String name =textViewFoto.getText().toString().replaceAll(" ","");
+                    String savedFile = saveImageFile(bitmap, "myFolder", name);
+
+                    File media = new File(savedFile);
+                    Uri imageUri =  Uri.fromFile(media);
+
+                    Intent share = new Intent(Intent.ACTION_SEND);
+                    share.setType("image/*");
+                    share.putExtra(Intent.EXTRA_STREAM, imageUri);
+                    share.putExtra(Intent.EXTRA_TITLE, textViewFoto.getText());
+                    startActivity(Intent.createChooser(share, "Share Image"));
+                }
+                catch (Throwable e) {
+                    Toast.makeText(getContext(), "Não foi possível executar a ação.", Toast.LENGTH_LONG).show();
+                }
             }
         });
 
-        photoViewModel.getLoading().observe(this, loading -> {
+        photoViewModel.getLoading().observe(getViewLifecycleOwner(), loading -> {
             if (loading) {
                 progressBar.setVisibility(View.VISIBLE);
             }
@@ -167,7 +196,10 @@ public class HomeFragment extends Fragment {
         SimpleDateFormat formatter = new SimpleDateFormat("yyyy-MM-dd");
         String todayString = formatter.format(currentTime);
         photoViewModel.getPhotoOfDay(todayString, API_KEY);
-        photoViewModel.liveData.observe(this, result -> {
+        photoViewModel.liveData.observe(getViewLifecycleOwner(), (PhotoResponse result) -> {
+
+            photoResponse = result;
+
             textViewFoto.setText(result.getTitle());
             if (result.getCopyright() != null) {
                 textViewAutor.setVisibility(View.VISIBLE);
@@ -183,149 +215,150 @@ public class HomeFragment extends Fragment {
         return view;
     }
 
+    @RequiresApi(api = Build.VERSION_CODES.O)
+    public static String saveImageFile(Bitmap image, String folder, String name){
+
+        boolean create = true;
+
+        File imageFile = new File(Environment.getExternalStorageDirectory() + "/" + folder);
+        if (!imageFile.exists()){
+            File screenShotsFolder = new File("/sdcard/Pictures/" + folder);
+            create = screenShotsFolder.mkdir();
+        }
+
+        File imageName = new File(new File("/sdcard/Pictures/" + folder + "/"), name + ".jpg");
+
+        try {
+            FileOutputStream outputStream = new FileOutputStream(imageName);
+            image.compress(Bitmap.CompressFormat.JPEG, 100, outputStream);
+            outputStream.flush();
+            outputStream.close();
+        }
+        catch (Throwable e){
+            e.printStackTrace();
+        }
+        return imageName.toPath().toString();
+    }
+
     private void updateLabel(Calendar myCalendar) {
-        Date[] dates = { myCalendar.getTime() };
         Date currentTime = myCalendar.getTime();
 
         String d = currentTime.toString();
         String[] currentDate = d.split(" ");
 
-        if (currentDate[1].equals("Jan")) {
-            textViewMes.setText("Jan");
-        }
-        else if (currentDate[1].equals("Feb")) {
-            textViewMes.setText("Fev");
-        }
-        else if (currentDate[1].equals("Mar")) {
-            textViewMes.setText("Mar");
-        }
-        else if (currentDate[1].equals("Apr")) {
-            textViewMes.setText("Abr");
-        }
-        else if (currentDate[1].equals("May")) {
-            textViewMes.setText("Mai");
-        }
-        else if (currentDate[1].equals("Jun")) {
-            textViewMes.setText("Jul");
-        }
-        else if (currentDate[1].equals("Jul")) {
-            textViewMes.setText("Jul");
-        }
-        else if (currentDate[1].equals("Aug")) {
-            textViewMes.setText("Ago");
-        }
-        else if (currentDate[1].equals("Sep")) {
-            textViewMes.setText("Set");
-        }
-        else if (currentDate[1].equals("Oct")) {
-            textViewMes.setText("Out");
-        }
-        else if (currentDate[1].equals("Nov")) {
-            textViewMes.setText("Nov");
-        }
-        else if (currentDate[1].equals("Dec")) {
-            textViewMes.setText("Dez");
+        switch (currentDate[1]) {
+            case "Jan":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jan));
+                break;
+            case "Feb":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.fev));
+                break;
+            case "Mar":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.mar));
+                break;
+            case "Apr":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.abr));
+                break;
+            case "May":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.mai));
+                break;
+            case "Jun":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jun));
+                break;
+            case "Jul":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jul));
+                break;
+            case "Aug":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.ago));
+                break;
+            case "Sep":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.set));
+                break;
+            case "Oct":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.out));
+                break;
+            case "Nov":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.nov));
+                break;
+            case "Dec":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.dez));
+                break;
         }
 
-        textViewAno.setText(currentDate[5].toString());
-        datePickerTimeline.deactivateDates(dates);
-        datePickerTimeline.setActiveDate(myCalendar);
+        textViewAno.setText(currentDate[5]);
     }
 
     private void changeDate() {
 
-        Date[] dates = { Calendar.getInstance().getTime() };
         Date currentTime = Calendar.getInstance().getTime();
 
         String d = currentTime.toString();
         String[] currentDate = d.split(" ");
 
-        if (currentDate[1].equals("Jan")) {
-            textViewMes.setText("Jan");
-        }
-        else if (currentDate[1].equals("Feb")) {
-            textViewMes.setText("Fev");
-        }
-        else if (currentDate[1].equals("Mar")) {
-            textViewMes.setText("Mar");
-        }
-        else if (currentDate[1].equals("Apr")) {
-            textViewMes.setText("Abr");
-        }
-        else if (currentDate[1].equals("May")) {
-            textViewMes.setText("Mai");
-        }
-        else if (currentDate[1].equals("Jun")) {
-            textViewMes.setText("Jul");
-        }
-        else if (currentDate[1].equals("Jul")) {
-            textViewMes.setText("Jul");
-        }
-        else if (currentDate[1].equals("Aug")) {
-            textViewMes.setText("Ago");
-        }
-        else if (currentDate[1].equals("Sep")) {
-            textViewMes.setText("Set");
-        }
-        else if (currentDate[1].equals("Out")) {
-            textViewMes.setText("Out");
-        }
-        else if (currentDate[1].equals("Nov")) {
-            textViewMes.setText("Nov");
-        }
-        else if (currentDate[1].equals("Dec")) {
-            textViewMes.setText("Dez");
-        }
-
-        final Calendar myCalendar = Calendar.getInstance();
-
-        textViewAno.setText(currentDate[5].toString());
-        datePickerTimeline.deactivateDates(dates);
-        datePickerTimeline.setActiveDate(myCalendar);
-        datePickerTimeline.setNextFocusUpId(800);
-    }
-
-    public void setTextDate(int month, int year) {
-        month++;
-        if (month == 1) {
-            textViewMes.setText("Jan");
-        }
-        else if (month == 2) {
-            textViewMes.setText("Fev");
-        }
-        else if (month == 3) {
-            textViewMes.setText("Mar");
-        }
-        else if (month == 4) {
-            textViewMes.setText("Abr");
-        }
-        else if (month == 5) {
-            textViewMes.setText("Mai");
-        }
-        else if (month == 6) {
-            textViewMes.setText("Jun");
-        }
-        else if (month == 7) {
-            textViewMes.setText("Jul");
-        }
-        else if (month == 8) {
-            textViewMes.setText("Ago");
-        }
-        else if (month == 9) {
-            textViewMes.setText("Set");
-        }
-        else if (month == 10) {
-            textViewMes.setText("Out");
-        }
-        else if (month == 11) {
-            textViewMes.setText("Nov");
-        }
-        else if (month == 12) {
-            textViewMes.setText("Dez");
+        switch (currentDate[1]) {
+            case "Jan":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jan));
+                break;
+            case "Feb":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.fev));
+                break;
+            case "Mar":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.mar));
+                break;
+            case "Apr":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.abr));
+                break;
+            case "May":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.mai));
+                break;
+            case "Jun":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jun));
+                break;
+            case "Jul":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.jul));
+                break;
+            case "Aug":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.ago));
+                break;
+            case "Sep":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.set));
+                break;
+            case "Out":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.out));
+                break;
+            case "Nov":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.nov));
+                break;
+            case "Dec":
+                textViewDia.setText(currentDate[2]);
+                textViewMes.setText(getString(R.string.dez));
+                break;
         }
 
-        String ano = String.valueOf(year);
-        textViewAno.setText(ano);
+        textViewAno.setText(currentDate[5]);
     }
 
     private void initViews(View view) {
@@ -333,66 +366,20 @@ public class HomeFragment extends Fragment {
         imageViewCalendar = view.findViewById(R.id.imageViewCalendar);
         imageViewUser = view.findViewById(R.id.imageViewUser);
 
+        textViewDia = view.findViewById(R.id.textViewDia);
         textViewMes = view.findViewById(R.id.textViewMes);
         textViewAno = view.findViewById(R.id.textViewAno);
 
-        datePickerTimeline = view.findViewById(R.id.datePickerTimeline);
-
-        datePickerTimeline.setDateTextColor(Color.WHITE);
-        datePickerTimeline.setDayTextColor(Color.WHITE);
-        datePickerTimeline.setMonthTextColor(Color.WHITE);
-        datePickerTimeline.setInitialDate(1995, 5, 18);
-
-        datePickerTimeline.setOnDateSelectedListener(new OnDateSelectedListener() {
-            @Override
-            public void onDateSelected(int year, int month, int day, int dayOfWeek) {
-                setTextDate(month, year);
-                String mes = "";
-                month++;
-                if (month == 1) {
-                    mes = "01";
-                }
-                else if (month == 2) {
-                    mes = "02";
-                }
-                else if (month == 3) {
-                    mes = "03";
-                }
-                else if (month == 4) {
-                    mes = "04";
-                }
-                else if (month == 5) {
-                    mes = "05";
-                }
-                else if (month == 6) {
-                    mes = "06";
-                }
-                else if (month == 7) {
-                    mes = "07";
-                }
-                else if (month == 8) {
-                    mes = "08";
-                }
-                else if (month == 9) {
-                    mes = "09";
-                }
-                dateRequest = year + "-" + mes + "-" + day;
-                getPhotoOfDay();
-            }
-
-            @Override
-            public void onDisabledDateSelected(int year, int month, int day, int dayOfWeek, boolean isDisabled) {
-
-            }
-        });
-
         photoViewModel = ViewModelProviders.of(this).get(PhotoViewModel.class);
+        photoResponse = new PhotoResponse();
 
         progressBar = view.findViewById(R.id.progress_bar);
 
         textViewFoto = view.findViewById(R.id.textViewFoto);
         textViewAutor = view.findViewById(R.id.textViewAutor);
         imageViewFoto = view.findViewById(R.id.imageViewFoto);
+        imageFavorite = view.findViewById(R.id.ic_favorite);
+        imageShare = view.findViewById(R.id.ic_share);
         textViewDescricao = view.findViewById(R.id.textViewDescricao);
         materialButtonPT = view.findViewById(R.id.materialButtonPT);
         materialButtonENG = view.findViewById(R.id.materialButtonENG);
@@ -401,11 +388,9 @@ public class HomeFragment extends Fragment {
 
     private void getPhotoOfDay () {
         photoViewModel.getPhotoOfDay(dateRequest, API_KEY);
-        photoViewModel.photo.observe(this, result -> {
+        photoViewModel.photo.observe(getViewLifecycleOwner(), result -> {
 
-//            if (!dateRequest.equals(result.getDate())) {
-//                Snackbar.make(container, R.string.nao_ha_imagem, Snackbar.LENGTH_LONG).show();
-//            }
+            photoResponse = result;
 
             textViewFoto.setText(result.getTitle());
             if (result.getCopyright() != null) {
@@ -420,155 +405,83 @@ public class HomeFragment extends Fragment {
         });
     }
 
-    public void getPortugueseDescription() {
+    private void getPortugueseDescription() {
         FirebaseTranslatorOptions options = new FirebaseTranslatorOptions.Builder().setSourceLanguage(FirebaseTranslateLanguage.EN).setTargetLanguage(FirebaseTranslateLanguage.PT).build();
         final FirebaseTranslator englishPortugueseTranslator = FirebaseNaturalLanguage.getInstance().getTranslator(options);
 
         FirebaseModelDownloadConditions conditions = new FirebaseModelDownloadConditions.Builder().requireWifi().build();
         englishPortugueseTranslator.downloadModelIfNeeded(conditions).addOnSuccessListener(
-                new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void v) {
-                        // Model downloaded successfully. Okay to start translating.
-                        // (Set a flag, unhide the translation UI, etc.)
-                        englishPortugueseTranslator.translate(textViewDescricao.getText().toString()).addOnSuccessListener(
-                                new OnSuccessListener<String>() {
-                                    @Override
-                                    public void onSuccess(@NonNull String translatedText) {
-                                        textViewDescricao.setText(translatedText);
-                                    }
-                                }).addOnFailureListener(
-                                new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        // Error.
-                                        // ...
-                                    }
-                                });
-                    }
+                v -> {
+
+                    englishPortugueseTranslator.translate(textViewDescricao.getText().toString()).addOnSuccessListener(
+                            translatedText -> textViewDescricao.setText(translatedText)).addOnFailureListener(
+                            e -> {
+
+                            });
                 }).addOnFailureListener(
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        // Model couldn’t be downloaded or other internal error.
-                        // ...
-                    }
+                e -> {
+
                 });
     }
 
-    public void getEnglishDescription() {
+    private void getEnglishDescription() {
         FirebaseTranslatorOptions options = new FirebaseTranslatorOptions.Builder().setSourceLanguage(FirebaseTranslateLanguage.PT).setTargetLanguage(FirebaseTranslateLanguage.EN).build();
         final FirebaseTranslator englishPortugueseTranslator = FirebaseNaturalLanguage.getInstance().getTranslator(options);
 
         FirebaseModelDownloadConditions conditions = new FirebaseModelDownloadConditions.Builder().requireWifi().build();
         englishPortugueseTranslator.downloadModelIfNeeded(conditions).addOnSuccessListener(
-                new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void v) {
-                        // Model downloaded successfully. Okay to start translating.
-                        // (Set a flag, unhide the translation UI, etc.)
-                        englishPortugueseTranslator.translate(textViewDescricao.getText().toString()).addOnSuccessListener(
-                                new OnSuccessListener<String>() {
-                                    @Override
-                                    public void onSuccess(@NonNull String translatedText) {
-                                        textViewDescricao.setText(translatedText);
-                                    }
-                                }).addOnFailureListener(
-                                new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        // Error.
-                                        // ...
-                                    }
-                                });
-                    }
+                v -> {
+
+                    englishPortugueseTranslator.translate(textViewDescricao.getText().toString()).addOnSuccessListener(
+                            translatedText -> textViewDescricao.setText(translatedText)).addOnFailureListener(
+                            e -> {
+
+                            });
                 }).addOnFailureListener(
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        // Model couldn’t be downloaded or other internal error.
-                        // ...
-                    }
+                e -> {
+
                 });
     }
 
-    public void getPortugueseTitle() {
+    private void getPortugueseTitle() {
         FirebaseTranslatorOptions options = new FirebaseTranslatorOptions.Builder().setSourceLanguage(FirebaseTranslateLanguage.EN).setTargetLanguage(FirebaseTranslateLanguage.PT).build();
         final FirebaseTranslator englishPortugueseTranslator = FirebaseNaturalLanguage.getInstance().getTranslator(options);
 
         FirebaseModelDownloadConditions conditions = new FirebaseModelDownloadConditions.Builder().requireWifi().build();
         englishPortugueseTranslator.downloadModelIfNeeded(conditions).addOnSuccessListener(
-                new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void v) {
-                        // Model downloaded successfully. Okay to start translating.
-                        // (Set a flag, unhide the translation UI, etc.)
-                        englishPortugueseTranslator.translate(textViewFoto.getText().toString()).addOnSuccessListener(
-                                new OnSuccessListener<String>() {
-                                    @Override
-                                    public void onSuccess(@NonNull String translatedText) {
-                                        textViewFoto.setText(translatedText);
-                                    }
-                                }).addOnFailureListener(
-                                new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        // Error.
-                                        // ...
-                                    }
-                                });
-                    }
+                v -> {
+
+                    englishPortugueseTranslator.translate(textViewFoto.getText().toString()).addOnSuccessListener(
+                            translatedText -> textViewFoto.setText(translatedText)).addOnFailureListener(
+                            e -> {
+
+                            });
                 }).addOnFailureListener(
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        // Model couldn’t be downloaded or other internal error.
-                        // ...
-                    }
+                e -> {
+
                 });
     }
 
-    public void getEnglishTitle() {
+    private void getEnglishTitle() {
         FirebaseTranslatorOptions options = new FirebaseTranslatorOptions.Builder().setSourceLanguage(FirebaseTranslateLanguage.PT).setTargetLanguage(FirebaseTranslateLanguage.EN).build();
         final FirebaseTranslator englishPortugueseTranslator = FirebaseNaturalLanguage.getInstance().getTranslator(options);
 
         FirebaseModelDownloadConditions conditions = new FirebaseModelDownloadConditions.Builder().requireWifi().build();
         englishPortugueseTranslator.downloadModelIfNeeded(conditions).addOnSuccessListener(
-                new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void v) {
-                        // Model downloaded successfully. Okay to start translating.
-                        // (Set a flag, unhide the translation UI, etc.)
-                        englishPortugueseTranslator.translate(textViewFoto.getText().toString()).addOnSuccessListener(
-                                new OnSuccessListener<String>() {
-                                    @Override
-                                    public void onSuccess(@NonNull String translatedText) {
-                                        textViewFoto.setText(translatedText);
-                                    }
-                                }).addOnFailureListener(
-                                new OnFailureListener() {
-                                    @Override
-                                    public void onFailure(@NonNull Exception e) {
-                                        // Error.
-                                        // ...
-                                    }
-                                });
-                    }
+                v -> {
+
+                    englishPortugueseTranslator.translate(textViewFoto.getText().toString()).addOnSuccessListener(
+                            translatedText -> textViewFoto.setText(translatedText)).addOnFailureListener(
+                            e -> {
+
+                            });
                 }).addOnFailureListener(
-                new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        // Model couldn’t be downloaded or other internal error.
-                        // ...
-                    }
+                e -> {
+
                 });
     }
 
-    private void replaceFragments(int container, Fragment fragment){
-        manager = getFragmentManager();
-        FragmentTransaction transaction = manager.beginTransaction()
-                .setCustomAnimations( R.anim.slide_up, 0, 0, R.anim.slide_down);
-        transaction.replace(container, fragment);
-        transaction.commit();
+    private void replaceFragments(Fragment fragment){
+        getFragmentManager().beginTransaction().setCustomAnimations(R.animator.slide_up, 0, 0, R.animator.slide_down).replace(R.id.fragment_container, fragment).commit();
     }
 }
