@@ -2,7 +2,7 @@ package com.example.spacenbeyond.view;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
-
+import android.app.Activity;
 import android.content.Intent;
 import android.graphics.Color;
 import android.graphics.Typeface;
@@ -18,16 +18,40 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 import android.widget.Toast;
-
 import com.example.spacenbeyond.R;
+import com.example.spacenbeyond.util.AppUtil;
+import com.facebook.CallbackManager;
+import com.facebook.FacebookCallback;
+import com.facebook.FacebookException;
+import com.facebook.login.LoginManager;
+import com.facebook.login.LoginResult;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.common.api.ApiException;
+import com.google.android.gms.tasks.Task;
+import com.google.android.material.snackbar.Snackbar;
 import com.google.android.material.textfield.TextInputLayout;
+import com.google.firebase.auth.FirebaseAuth;
+import com.facebook.FacebookSdk;
+import com.facebook.appevents.AppEventsLogger;
+
+import java.util.Arrays;
 
 public class LoginActivity extends AppCompatActivity {
 
     private TextView txtCadastro;
     private Button btnLogin;
+    private Button gButton;
+    private Button fButton;
     private TextInputLayout txtEmail;
     private TextInputLayout txtSenha;
+    private GoogleSignInClient googleSignInClient;
+    private CallbackManager callbackManager;
+
+    public static final String GOOGLE_ACCOUNT = "google_account";
+    public static int RC_SIGN_IN = 101;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -38,36 +62,111 @@ public class LoginActivity extends AppCompatActivity {
 
         textoClicavel();
 
-        btnLogin.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-               verifyFields();
-            }
+        btnLogin.setOnClickListener(v -> verifyFields());
+
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestEmail()
+                .build();
+
+        googleSignInClient = GoogleSignIn.getClient(this, gso);
+
+        gButton.setOnClickListener(view -> {
+            Intent signInIntent = googleSignInClient.getSignInIntent();
+            startActivityForResult(signInIntent, RC_SIGN_IN);
         });
 
+        fButton.setOnClickListener(view -> loginFacebook());
     }
 
-    public void verifyFields() {
+    @Override
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
+        callbackManager.onActivityResult(requestCode, resultCode, data);
+        super.onActivityResult(requestCode, resultCode, data);
+        if (resultCode == Activity.RESULT_OK)
+            if (requestCode == RC_SIGN_IN) {
+                try {
+                    Task<GoogleSignInAccount> task = GoogleSignIn.getSignedInAccountFromIntent(data);
+                    GoogleSignInAccount account = task.getResult(ApiException.class);
+                    onLoggedIn(account);
+                } catch (ApiException e) {
+                    Toast.makeText(getApplicationContext(), "Não foi possível fazer o login", Toast.LENGTH_SHORT).show();
+                }
+            }
+    }
+
+    private void onLoggedIn(GoogleSignInAccount googleSignInAccount) {
+        Intent intent = new Intent(this, MainActivity.class);
+        intent.putExtra(GOOGLE_ACCOUNT, googleSignInAccount);
+        startActivity(intent);
+        finish();
+    }
+
+    private void loginFacebook(){
+        LoginManager.getInstance().logInWithReadPermissions(this, Arrays.asList("public_profile"));
+
+        LoginManager.getInstance().registerCallback(callbackManager, new FacebookCallback<LoginResult>() {
+            @Override
+            public void onSuccess(LoginResult loginResult) {
+                irParaHome(loginResult.getAccessToken().getUserId());
+            }
+
+            @Override
+            public void onCancel() {
+                Toast.makeText(getApplicationContext(),"Cancelado pelo usuário", Toast.LENGTH_LONG).show();
+
+            }
+
+            @Override
+            public void onError(FacebookException error) {
+                Toast.makeText(getApplicationContext(),"Erro ao logar com Facebook", Toast.LENGTH_LONG).show();
+
+            }
+        });
+    }
+
+    private void irParaHome(String uiid){
+        AppUtil.salvarIdUsuario(this, uiid);
+        Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+        startActivity(intent);
+    }
+
+    private void verifyFields() {
         String email = txtEmail.getEditText().getText().toString();
         String senha = txtSenha.getEditText().getText().toString();
 
         if (!email.isEmpty() && !senha.isEmpty()) {
-            Intent intent = new Intent(LoginActivity.this, MainActivity.class);
-            startActivity(intent);
+
+            FirebaseAuth.getInstance().signInWithEmailAndPassword(email, senha)
+                    .addOnCompleteListener(task -> {
+                        // Caso login com sucesso vamos para tela  Home
+                        if (task.isSuccessful()) {
+                            AppUtil.salvarIdUsuario(getApplicationContext(), FirebaseAuth.getInstance().getCurrentUser().getUid());
+
+                            Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+                            startActivity(intent);
+                        }
+                        else {
+                            // Se deu algum erro mostramos para o usuário a mensagem
+                            Snackbar.make(btnLogin, "Erro ao tentar logar \n" + task.getException().getMessage(), Snackbar.LENGTH_SHORT).show();
+                        }
+                    });
+
         }
         else {
             Toast.makeText(LoginActivity.this, "Por favor, forneça os dados necessários para login.", Toast.LENGTH_LONG).show();
         }
     }
 
-    public void initViews() {
+    private void initViews() {
 
         btnLogin = findViewById(R.id.btnLogin);
         txtCadastro = findViewById(R.id.textViewCadastreSe);
         txtEmail = findViewById(R.id.txtEmail);
         txtSenha = findViewById(R.id.txtSenha);
+        gButton = findViewById(R.id.googleButton);
+        fButton = findViewById(R.id.facebookButton);
+        callbackManager = CallbackManager.Factory.create();
     }
-
 
     private void textoClicavel() {
         String text = "Ainda não tem conta? Cadastre-se.";
