@@ -1,7 +1,6 @@
 package com.example.spacenbeyond.data.remote;
 
 import com.example.spacenbeyond.model.PhotoResponse;
-
 import io.reactivex.Observable;
 import retrofit2.http.GET;
 import retrofit2.http.Query;
@@ -9,5 +8,8 @@ import retrofit2.http.Query;
 public interface PhotoAPI {
 
     @GET("planetary/apod")
-    Observable<PhotoResponse> getPhotoOfDay(@Query("date") String date, @Query("api_key") String apiKey);
+    Observable<PhotoResponse> getPhotoOfDay(
+            @Query("date") String date,
+            @Query("api_key") String apiKey
+    );
 }
