@@ -1,4 +1,3 @@
-
 package com.example.spacenbeyond.model;
 
 import com.google.gson.annotations.SerializedName;
@@ -7,18 +6,25 @@ public class PhotoResponse {
 
     @SerializedName("copyright")
     private String mCopyright;
+
     @SerializedName("date")
     private String mDate;
+
     @SerializedName("explanation")
     private String mExplanation;
+
     @SerializedName("hdurl")
     private String mHdurl;
+
     @SerializedName("media_type")
     private String mMediaType;
+
     @SerializedName("service_version")
     private String mServiceVersion;
+
     @SerializedName("title")
     private String mTitle;
+
     @SerializedName("url")
     private String mUrl;
 
@@ -26,63 +32,24 @@ public class PhotoResponse {
         return mCopyright;
     }
 
-    public void setCopyright(String copyright) {
-        mCopyright = copyright;
-    }
+    public String getDate() { return mDate; }
 
-    public String getDate() {
-        return mDate;
-    }
-
-    public void setDate(String date) {
-        mDate = date;
-    }
 
     public String getExplanation() {
         return mExplanation;
     }
 
-    public void setExplanation(String explanation) {
-        mExplanation = explanation;
-    }
-
-    public String getHdurl() {
-        return mHdurl;
-    }
-
-    public void setHdurl(String hdurl) {
-        mHdurl = hdurl;
-    }
-
-    public String getMediaType() {
-        return mMediaType;
-    }
-
-    public void setMediaType(String mediaType) {
-        mMediaType = mediaType;
-    }
-
-    public String getServiceVersion() {
-        return mServiceVersion;
-    }
-
-    public void setServiceVersion(String serviceVersion) {
-        mServiceVersion = serviceVersion;
-    }
 
     public String getTitle() {
         return mTitle;
     }
 
-    public void setTitle(String title) {
-        mTitle = title;
-    }
 
     public String getUrl() {
         return mUrl;
     }
 
-    public void setUrl(String url) {
-        mUrl = url;
+    public void setmCopyright(String mCopyright) {
+        this.mCopyright = mCopyright;
     }
 }
