@@ -1,18 +1,19 @@
 package com.example.spacenbeyond.view;
 
-import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import androidx.fragment.app.Fragment;
-
+import android.Manifest;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.view.MenuItem;
-
 import com.example.spacenbeyond.R;
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
 public class MainActivity extends AppCompatActivity {
 
-    private BottomNavigationView bottomNav;
+    public BottomNavigationView bottomNav;
+    private static final int PERMISSION_CODE = 100;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -21,34 +22,50 @@ public class MainActivity extends AppCompatActivity {
 
         bottomNav = findViewById(R.id.bottom_navigation);
 
+        captureImage();
+
         bottomNav.setOnNavigationItemSelectedListener(navListener);
         getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, new HomeFragment()).commit();
     }
 
-    private BottomNavigationView.OnNavigationItemSelectedListener navListener = new BottomNavigationView.OnNavigationItemSelectedListener() {
-        @Override
-        public boolean onNavigationItemSelected(@NonNull MenuItem item) {
-            Fragment selectedFragment = null;
+    private void captureImage() {
+        int permissionCamera = ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.CAMERA);
+        int permissionStorage = ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.WRITE_EXTERNAL_STORAGE);
+        int permissionRead = ContextCompat.checkSelfPermission(getApplicationContext(), Manifest.permission.READ_EXTERNAL_STORAGE);
 
-            switch (item.getItemId()){
-                case R.id.nav_home:
-                    selectedFragment = new HomeFragment();
-                    break;
-                case R.id.nav_hashtag:
-                    selectedFragment = new HashtagFragment();
-                    break;
-                case R.id.nav_help:
-                    selectedFragment = new AjudaFragment();
-                    break;
-            }
-
-            getSupportFragmentManager().beginTransaction().replace(R.id.fragment_container, selectedFragment).commit();
-            return true;
+        if (permissionCamera == PackageManager.PERMISSION_GRANTED && permissionStorage == PackageManager.PERMISSION_GRANTED && permissionRead == PackageManager.PERMISSION_GRANTED) { }
+        else {
+            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA, Manifest.permission.WRITE_EXTERNAL_STORAGE, Manifest.permission.READ_EXTERNAL_STORAGE}, PERMISSION_CODE);
         }
+    }
+
+    private final BottomNavigationView.OnNavigationItemSelectedListener navListener = item -> {
+        Fragment selectedFragment = null;
+
+        switch (item.getItemId()) {
+            case R.id.nav_home:
+                selectedFragment = new HomeFragment();
+                item.setChecked(true);
+                break;
+            case R.id.nav_hashtag:
+                selectedFragment = new HashtagFragment();
+                break;
+            case R.id.nav_favorites:
+                selectedFragment = new FavoritosFragment();
+                break;
+            case R.id.nav_help:
+                selectedFragment = new AjudaFragment();
+                break;
+        }
+
+        replaceFragments(selectedFragment);
+        return true;
     };
 
-    @Override
-    public void onBackPressed() {
-            super.onBackPressed();
+    private void replaceFragments(Fragment fragment) {
+        getSupportFragmentManager().beginTransaction().setCustomAnimations(R.animator.slide_up, 0, 0, 0).replace(R.id.fragment_container, fragment).commit();
     }
+
+    @Override
+    protected void onResume() { super.onResume(); }
 }
